Change checkbox field with normal checkbox for remember checkbox
